final-pjt-back/requirements.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
PyJWT==1.7.1
pyparsing==3.0.6
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.7
